[refactor] 대기열 생성, 순번 조회 : db > redis 전환

+++ src/main/java/hhplus/tdd/concert/app/api/dto/response/waiting/WaitingNumRes.java
@@ -3,7 +3,7 @@ package hhplus.tdd.concert.app.api.dto.response.waiting;
 import hhplus.tdd.concert.app.application.waiting.dto.WaitingNumQuery;
 
 public record WaitingNumRes(
-        int num
+        Long num
 ) {
 
     public static WaitingNumRes from(WaitingNumQuery dto) {
