+++ src/main/java/hhplus/tdd/concert/app/application/waiting/dto/WaitingNumQuery.java
@@ -1,5 +1,5 @@
 package hhplus.tdd.concert.app.application.waiting.dto;
 
 public record WaitingNumQuery(
-        int num
+        Long num
 ) {}
